+++ src/main/java/insane96mcp/enhancedai/modules/mobs/anticheese/AntiCheese.java
@@ -3,7 +3,6 @@ package insane96mcp.enhancedai.modules.mobs.anticheese;
 import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
@@ -13,24 +12,20 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.monster.Enemy;
-import net.minecraft.world.entity.vehicle.Boat;
-import net.minecraft.world.entity.vehicle.Minecart;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.EntityMountEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Anti-Cheese", description = "Prevent players from abusing some game mechanics to stop mobs. Only mobs in the entity type tag enhancedai:can_use_anti_cheese will be affected by this feature.")
-@LoadFeature(module = Modules.Ids.MOBS)
+@LoadFeature(module = Modules.Ids.MOBS, name = "Anti-Cheese", description = "Prevent players from abusing some game mechanics to stop mobs. Only mobs in the entity type tag enhancedai:can_use_anti_cheese will be affected by this feature. Only veichles in `enhancedai:anti_cheese_veichles` will be affected by this feature.")
 public class AntiCheese extends Feature {
     public static final TagKey<EntityType<?>> CAN_USE_ANTI_CHEESE = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "can_use_anti_cheese"));
+    public static final TagKey<EntityType<?>> ANTI_CHEESE_VEICHLES = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "anti_cheese_vehicles"));
 
-    @Config
-    @Label(name = "Prevent Boating & Minecarting", description = "If true, 'Enemies' will no longer be able to be Boated and Minecarted.")
-    public static Boolean preventBoating = false;
+    @Config(description = "If true, 'Enemies' will no longer be able to be put in vehicles.")
+    public static Boolean preventRidingVehicles = false;
 
-    @Config
-    @Label(name = "Break trapping vehicles", description = "If true, 'Enemies' will break boats or minecarts if boated or minecarted.")
-    public static Boolean antiBoatAndMinecart = true;
+    @Config(description = "If true, 'Enemies' will break vehicles.")
+    public static Boolean breakVehicles = true;
 
     public AntiCheese(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
@@ -40,22 +35,22 @@ public class AntiCheese extends Feature {
     public void onMount(EntityMountEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntityMounting() instanceof Enemy)
-                || !(event.getEntityBeingMounted() instanceof Boat) && !(event.getEntityBeingMounted() instanceof Minecart)
-                || !event.getEntityMounting().getType().is(CAN_USE_ANTI_CHEESE))
+                || !(event.getEntityBeingMounted().getType().is(ANTI_CHEESE_VEICHLES))
+                || !event.getEntityMounting().getType().is(CAN_USE_ANTI_CHEESE)
+                || !preventRidingVehicles)
             return;
 
-        if (preventBoating)
-            event.setCanceled(true);
+        event.setCanceled(true);
     }
 
     @SubscribeEvent
     public void onJoinLevel(EntityJoinLevelEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntity() instanceof Mob mob)
-                || !mob.getType().is(CAN_USE_ANTI_CHEESE))
+                || !mob.getType().is(CAN_USE_ANTI_CHEESE)
+                || !breakVehicles)
             return;
 
-        if (antiBoatAndMinecart)
-            mob.goalSelector.addGoal(1, new BreakVehicleGoal(mob));
+        mob.goalSelector.addGoal(1, new BreakVehicleGoal(mob));
     }
 }